Start breaking up files
