+++ metrics/mypy_high_water_mark
@@ -1,1 +1,1 @@
-91.8900
+91.9000
